condensed api routing and fixed stalled requests error  via external resolver
