+++ tsconfig.json
@@ -21,8 +21,6 @@
     "next-env.d.ts",
     "**/*.ts",
     "**/*.tsx",
-    "../wrk_lua/execScripts.ts",
-    "../wrk_lua/luaFileGenerator.js"
-  ],
+    "methods/getHistogramData.js"  ],
   "exclude": ["node_modules"]
 }
